3/22 push/pull start

+++ src/main/java/com/ts/mvc/module/status/PetStatus.java
@@ -28,7 +28,7 @@ public class PetStatus {
 
 	@Id
 	@GeneratedValue
-	private Integer statusIdx;
+	private Long statusIdx;
 	
 	private String chosenDate;
